backend/requirements.txt
# =============================================================================
# NetOps-Flow Backend Dependencies
# =============================================================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Async Task Queue
celery==5.3.6
redis==5.0.1

# Validation & Configuration
pydantic==2.5.3
pydantic-settings==2.1.0

# Security
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
cryptography==41.0.7

# Remote Execution
paramiko==3.4.0
pywinrm==0.4.3
python-nmap==0.7.1

# Docker SDK (for sandboxing)
docker==7.0.0

# Logging
python-json-logger==2.0.7
